requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
cattrs==1.10.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cytoolz==0.11.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-allauth==0.51.0
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-countries==7.3.2
django-extensions==3.1.5
django-import-export==2.8.0
django-nested-admin==3.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-s3direct==2.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
etherscan==0.3.0
frozenlist==1.3.0
gcloud==0.17.0
googleapis-common-protos==1.56.2
hexbytes==0.2.2
httplib2==0.20.4
idna==3.3
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.0
jws==0.1.3
lru-dict==1.1.7
MarkupPy==1.14
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauth2client==3.0.0
oauthlib==3.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
parsimonious==0.8.1
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==1.7.1
pyparsing==3.0.9
Pyrebase==3.0.27
pyrsistent==0.18.1
python-jwt==2.0.1
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.11.1
requests-cache==0.9.4
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rlp==2.0.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toolz==0.11.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.9
varint==1.0.2
web3==5.29.2
websockets==9.1
Werkzeug==2.1.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
